Refactor Apollo Client code into its own file and folder
