+++ src/graphql/apollo/apolloClient.ts
@@ -1,0 +1,31 @@
+import { ApolloClient, HttpLink, InMemoryCache, split } from "@apollo/client";
+import { GraphQLWsLink } from "@apollo/client/link/subscriptions";
+import { createClient } from "graphql-ws";
+import { getMainDefinition } from "@apollo/client/utilities";
+
+const httpLink = new HttpLink({
+  uri: `${process.env.REACT_APP_BACKEND_DOMAIN}/graphql`,
+});
+
+const wsLink = new GraphQLWsLink(
+  createClient({
+    url: process.env.REACT_APP_BACKEND_WEBSOCKET!,
+  })
+);
+
+const splitLink = split(
+  ({ query }) => {
+    const definition = getMainDefinition(query);
+    return (
+      definition.kind === "OperationDefinition" &&
+      definition.operation === "subscription"
+    );
+  },
+  wsLink,
+  httpLink
+);
+
+export const apolloClient = new ApolloClient({
+  link: splitLink,
+  cache: new InMemoryCache(),
+});
