+++ src/index.tsx
@@ -4,47 +4,12 @@ import App from "./App";
 import { store } from "./redux/store";
 import { Provider } from "react-redux";
 import * as serviceWorkerRegistration from "./serviceWorkerRegistration";
-import {
-  ApolloClient,
-  ApolloProvider,
-  HttpLink,
-  InMemoryCache,
-  split,
-} from "@apollo/client";
-import { GraphQLWsLink } from "@apollo/client/link/subscriptions";
-import { createClient } from "graphql-ws";
-import { getMainDefinition } from "@apollo/client/utilities";
+import { ApolloProvider } from "@apollo/client";
+import { apolloClient } from "./graphql/apollo/apolloClient";
 
 const root = ReactDOM.createRoot(
   document.getElementById("root") as HTMLElement
 );
-
-const httpLink = new HttpLink({
-  uri: `${process.env.REACT_APP_BACKEND_DOMAIN}/graphql`,
-});
-
-const wsLink = new GraphQLWsLink(
-  createClient({
-    url: process.env.REACT_APP_BACKEND_WEBSOCKET!,
-  })
-);
-
-const splitLink = split(
-  ({ query }) => {
-    const definition = getMainDefinition(query);
-    return (
-      definition.kind === "OperationDefinition" &&
-      definition.operation === "subscription"
-    );
-  },
-  wsLink,
-  httpLink
-);
-
-const apolloClient = new ApolloClient({
-  link: splitLink,
-  cache: new InMemoryCache(),
-});
 
 root.render(
   <React.StrictMode>
